% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armaString.R
\name{armaString}
\alias{armaString}
\title{armaString}
\usage{
armaString(model, round_to = 2)
}
\arguments{
\item{model}{arma/arima model to write as a string}

\item{round_to}{number of decimal places to keep}
}
\value{
a string
}
\description{
This function outputs a string of the arma model's equation
}
\examples{
series <- c(1, 3, 2, 3, 4, 3, 4, 5, 6, 4, 5, 6, 5, 4, 7, 5, 4, 7, 8, 5, 7, 8, 9, 7, 9)
arimamodel <- forecast::auto.arima(series)
armaString(arimamodel)

}
