% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindOutliers.R
\name{FindOutliers}
\alias{FindOutliers}
\title{FindOutliers}
\usage{
FindOutliers(x, IQRs = 3, plot = FALSE)
}
\arguments{
\item{x}{a numeric vector or time series to be processed}

\item{IQRs}{Allowable number of IQRs' away from the first or third quartile for a point to not be considered an outlier. Higher numbers result in fewer outliers.}

\item{plot}{Enables a plot highlighting detected outliers}
}
\value{
A vector with likely weights of outliers.
}
\description{
This function identifies time series outliers in a data signal
}
\examples{
FindOutliers(c(2,1,2,2,0,2,20000,2,2,1,2))
FindOutliers(c(1,2,3,5,-10,4,2,4,2,6,8,9), plot=TRUE)

}
