% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSeriesToCutoff.R
\name{GenerateSeriesToCutoff}
\alias{GenerateSeriesToCutoff}
\title{GenerateSeriesToCutoff}
\usage{
GenerateSeriesToCutoff(N, mu, sigma, cutoff_point, starting_point)
}
\arguments{
\item{N}{Maximum length of series}

\item{mu}{mean}

\item{sigma}{variance}

\item{cutoff_point}{threshold at which to cut series}

\item{starting_point}{first point in generated series}
}
\value{
A vector of points
}
\description{
This function generates a random series with a set trend, and will cut the series as soon as it crosses a certain threshold. The series, p, is generated as follows: p[i - 1] * exp(rnorm(1, mu, sigma)
}
\examples{
GenerateSeriesToCutoff(100, .01, .005, 200, 100)
GenerateSeriesToCutoff(100, .1, .001, 100, 50)

}
