% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installnload_packages.R
\name{installnload_packages}
\alias{installnload_packages}
\title{installnload_packages}
\usage{
installnload_packages(packages)
}
\arguments{
\item{packages}{a character vector of package names to be loaded}
}
\description{
This function checks to see if packages are installed, installs them if needed, and then loads and attaches them. This can be useful in reproducable work, when one uses a more obscure package. Granted, containing it in a package creates another problem, but the code can be distributed and copied easily.
}
