% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printClassificationTable.R
\name{printClassificationTable}
\alias{printClassificationTable}
\title{printClassificationTable}
\usage{
printClassificationTable(observed, predicted, method)
}
\arguments{
\item{observed}{Vector of actual values}

\item{predicted}{Vector of predicted values}

\item{method}{A string to pass into the output, identifies the model used.}
}
\value{
Prints a classification Matrix and calculates the misclassification rate.
}
\description{
printClassificationTable
}
\examples{
printClassificationTable(c(1,2,1,2,1), c(1,2,2,1,1), method="input")
}
