% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prognosis_ARIMA.R
\name{Prognosis_ARIMA}
\alias{Prognosis_ARIMA}
\title{Prognosis_ARIMA}
\usage{
Prognosis_ARIMA(data, maxp = 5, maxq = 5, maxd = 2, tendtomean = FALSE,
  forecastLength = NULL, forecastConfidence = 50, outlierTolerance = 2,
  detailed = FALSE)
}
\arguments{
\item{data}{Input data, as a vector}

\item{maxp}{Maximum number of AR terms}

\item{maxq}{Maximum number of MA terms}

\item{maxd}{Maximum number of differencing}

\item{tendtomean}{Chooses whether or not to include a mean in the ARIMA model}

\item{forecastLength}{number of points to forecast - default is .2 of the input's length}

\item{forecastConfidence}{Level for plotted and output confidence interval}

\item{outlierTolerance}{Allowable number of IQRs' away from the first or third quartile for a point to not be considered an outlier. Higher numbers result in fewer outliers.}

\item{detailed}{Allows for more detailed output, where applicable. Defaults to FALSE}
}
\value{
A list consisting of a plot, a model, and predictions
}
\description{
This function predicts future values using {\href{https://en.wikipedia.org/wiki/Autoregressive_integrated_moving_average}{ARIMA}} and {\href{https://en.wikipedia.org/wiki/Autoregressive_conditional_heteroskedasticity#GARCH}{GARCH}} models.
}
\examples{
exampledata <- GenerateSeriesToCutoff(100, .01, .005, 200, 100)
Prognosis_output <- Prognosis_ARIMA(exampledata)
Prognosis_output$Plot
Prognosis_output$Model
Prognosis_output$Table
}
