% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prognosis_Seasonal.R
\name{Prognosis_Seasonal}
\alias{Prognosis_Seasonal}
\title{Prognosis_Seasonal}
\usage{
Prognosis_Seasonal(data, times, seasonality, forecastLength = NULL,
  forecastConfidence = 50, outlierTolerance = 2, detailed = FALSE)
}
\arguments{
\item{data}{Input data, as a vector}

\item{times}{Vector of times for each data point}

\item{seasonality}{How many data point there are in each seasonal period}

\item{forecastLength}{Number of points to forecast - default is .2 of the input's length}

\item{forecastConfidence}{Level for plotted and output confidence interval}

\item{outlierTolerance}{Allowable number of IQRs' away from the first or third quartile for a point to not be considered an outlier. Higher numbers result in fewer outliers.}

\item{detailed}{Allows for more detailed output, where applicable. Defaults to FALSE}
}
\value{
A list consisting of a plot, a model, and predictions
}
\description{
Prognosis_Seasonal
}
\examples{
exampledata <- GenerateSeriesToCutoff(100, .01, .005, 200, 100)
times <- seq(as.Date("2000/1/1"), by = "month", length.out = length(exampledata))
Prognosis_output<- Prognosis_Seasonal(exampledata, times, 12)
Prognosis_output$Plot
Prognosis_output$Model
Prognosis_output$Table

}
