% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_png.R
\name{save_as_png}
\alias{save_as_png}
\title{save_as_png}
\usage{
save_as_png(plotme, file, width = 512, height = 512, axes = T)
}
\arguments{
\item{plotme}{plot to be saved}

\item{file}{path at which plot will be saved}

\item{width}{im pixels}

\item{height}{in pixels}

\item{axes}{boolean that controls axes}
}
\value{
generates an image file at the designated path
}
\description{
save_as_png
}
