% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_GR.R
\name{theme_GR}
\alias{theme_GR}
\title{theme_GR}
\usage{
theme_GR()
}
\arguments{
\item{theme}{Select one element from the following list}
}
\value{
the ggplot2 function theme_manual(), with chosen changes
}
\description{
theme_GR
}
\examples{
d <- ggplot2::qplot(carat, data = ggplot2::diamonds[ggplot2::diamonds$color \%in\%LETTERS[4:7], ], geom = "histogram", bins=30, fill = color)
d + theme_GR()

}
