% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesAnalysis.R
\name{TimeSeriesAnalysis}
\alias{TimeSeriesAnalysis}
\title{TimeSeriesAnalysis}
\usage{
TimeSeriesAnalysis(valuecol, datecol, plot = TRUE, plotly = FALSE,
  IQRs = 2)
}
\arguments{
\item{valuecol}{Data vector}

\item{datecol}{Date vector}

\item{plot}{toggles drawing of plot}

\item{plotly}{toggles plotting in plotly or ggplot}

\item{IQRs}{Allowable number of IQRs' away from the first or third quartile for a point to not be considered an outlier. Higher numbers result in fewer outliers.}
}
\value{
a data frame with the times, data values, outliers, changes in mean, and changes in variance
}
\description{
Detects and marks outliers, breakpoints, and changes in variance in a time series.
}
\examples{
data <- c(rnorm(50,0,1), rnorm(50,0,2),20,rnorm(50,10,1))
dates <- seq(as.Date("2016/1/1"), by="days", length.out=151)
TimeSeriesAnalysis(data,dates)

TimeSeriesAnalysis(c(1,2,3,5,-10,4,2,4,2,6,8,9), seq(as.Date("2016/1/1"), by="days", length.out=12))
}
