% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnsupervisedClusters.R
\name{UnsupervisedClusters}
\alias{UnsupervisedClusters}
\title{UnsupervisedClusters}
\usage{
UnsupervisedClusters(datain, plottype = "fan", forceclusters = NULL,
  classification = NULL)
}
\arguments{
\item{datain}{a dataframe which will be clustered}

\item{plottype}{one of the following: "dendogram", "fan", "radial","unrooted","cladogram"}

\item{forceclusters}{allows for specification of number of clusters. If NULL, the best number will be calculated.}

\item{classification}{optional - a vector with true class values. If given, a confusion matrix will be output}
}
\value{
datain, but with an added column with clusters
}
\description{
This function performs unsupervised clustering on a dataset. This technique can be useful to detect hidden trends in the data, as well as outliers and odd behaviors.
Works best with smaller datasets - more than a few thousand can slow greatly and be less useful
}
\examples{
set.seed(1)
datain <- data.frame(matrix(rnorm(400), nrow=100))

dataout <- UnsupervisedClusters(datain)
dataout$Plot

dataout$index <- 1:length(dataout$Data[,1])
graphme <- reshape::melt(dataout, id=c("index", "cluster"))
ggplot2::ggplot(graphme, ggplot2::aes(x=index, y=value, color=cluster)) +
 ggplot2::geom_point() + ggplot2::facet_wrap(~variable) + theme_GR()
}
