% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_explore.R
\name{bayes_explore}
\alias{bayes_explore}
\title{bayes_explore}
\usage{
bayes_explore(model, shiny = F, show_intercept = T)
}
\arguments{
\item{model}{a model, in lm or glm format}

\item{shiny}{boolean, True opens shinystan interface while False outputs a handful of plots}

\item{show_intercept}{boolean. Toggles if the intercept is included in plots}
}
\value{
a model and either a shiny interface, or handful of plots
}
\description{
This function uses stan to allow the user to make a bayesian model from a frequentist one, allowing for easy bayesian conclusions (intervals, estimates, etc.)
There is an option to open a shiny-based analysis portal, or to output some plots with posterior distributions and population estimates.
The model itself is also output.
}
\examples{
bayes_explore(lm(mpg~., mtcars))
bayes_explore(lm(stack.loss~., stackloss), show_intercept=FALSE)
bayes_explore(lm(Fertility~0+., swiss))
bayes_explore(glm(am~0+mpg+wt, mtcars, family=binomial(link='logit')))

}
