% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelString.R
\name{modelString}
\alias{modelString}
\title{modelString}
\usage{
modelString(model, round_to = 2, dependent = "y")
}
\arguments{
\item{model}{model to write out}

\item{round_to}{integer: number of significant digits to round to}

\item{dependent}{string: the dependent variable. Defaults to y}
}
\value{
a latex formatted string
}
\description{
Writes a model's equation in LaTeX
}
\examples{
model <- lm(mpg~.,data=mtcars)
modelString(model)

}
